new temp

+++ ProntoV2/ProntoV2/Item.cs
@@ -17,7 +17,7 @@ namespace ProntoV2
 
 
         public Item() { }
-        public Item(string ic, bool it, string name,string mn,string mCountry, string Description,string unitQty, float qty,bool bIsWhighted, bool quantitInPackage,string unitOfMeasure, int price,float unitofmeasureprice, bool allowD, bool status)
+        public Item(string ic, bool it, string name,string mn,string mCountry, string Description,string unitQty, double qty,bool bIsWhighted, bool quantitInPackage,string unitOfMeasure, double price,double unitofmeasureprice, bool allowD, bool status)
         {
             ItemCode = ic;
             ItemType = it;
@@ -45,12 +45,12 @@ namespace ProntoV2
         public string ManufactureCountry { get; set; }
         public string ManufacturerItemDescription { get; set; }
         public string UnitQty { get; set; }
-        public float Quantity { get; set; }
+        public double Quantity { get; set; }
         public bool bIsWeighted { get; set; }
         public string UnitOfMeasure { get; set; }
         public bool QtyInPackage { get; set; }
-        public int ItemPrice { get; set; }
-        public float UnitOfMeasurePrice { get; set; }
+        public double ItemPrice { get; set; }
+        public double UnitOfMeasurePrice { get; set; }
         public bool AllowDiscount { get; set; }
         public bool ItemStatus { get; set; }
     }
